requirements.txt
flask==3.0.3
requests==2.32.3
PyYAML==6.0.2